♻️ Refactor staking functions into composable
